Отделит шаблоны от классов
