+++ src/view/board-view/template.js
@@ -1,0 +1,7 @@
+const createBoardTemplate = () => `
+  <section class="trip-events">
+    <h2 class="visually-hidden">Trip events</h2>
+  </section>
+`;
+
+export default createBoardTemplate;
