--- src/view/new-point-button-view.js
@@ -1,10 +1,0 @@
-import AbstractView from '../framework/view/abstract-view';
-
-const createNewPointButtonTemplate = () =>
-  '<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>';
-
-export default class NewPointButtonView extends AbstractView {
-  get template() {
-    return createNewPointButtonTemplate();
-  }
-}
