+++ src/view/new-point-button-view/new-point-button-view.js
@@ -1,0 +1,8 @@
+import AbstractView from '../../framework/view/abstract-view';
+import createNewPointButtonTemplate from './template';
+
+export default class NewPointButtonView extends AbstractView {
+  get template() {
+    return createNewPointButtonTemplate();
+  }
+}
