+++ src/view/new-point-button-view/template.js
@@ -1,0 +1,4 @@
+const createNewPointButtonTemplate = () =>
+  '<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>';
+
+export default createNewPointButtonTemplate;
