+++ src/view/point-list-view/point-list-view.js
@@ -1,0 +1,8 @@
+import AbstractView from '../../framework/view/abstract-view.js';
+import createPointListTemplate from './template.js';
+
+export default class PointListView extends AbstractView {
+  get template() {
+    return createPointListTemplate();
+  }
+}
