+++ src/view/point-list-view/template.js
@@ -1,0 +1,3 @@
+const createPointListTemplate = () => '<ul class="trip-events__list"></ul>';
+
+export default createPointListTemplate;
