+++ src/view/point-view/template.js
@@ -1,11 +1,10 @@
-import { DateFormat } from '../const.js';
-import AbstractView from '../framework/view/abstract-view.js';
+import { DateFormat } from '../../const.js';
 import {
   getDestinationById,
   getSelectedOffersByType,
   humanizeDateFormat,
   humanizeDurationEvent,
-} from '../utils/utils.js';
+} from '../../utils/utils.js';
 
 const createOfferItemTemplate = ({ title, price }) => `
   <li class="event__offer">
@@ -106,35 +105,4 @@ const createPointTemplate = ({ point, destinations, offers }) => {
   `;
 };
 
-export default class PointView extends AbstractView {
-  #point = [];
-  #destinations = [];
-  #offers = [];
-
-  #handleEditClick = null;
-
-  constructor({ point, destinations, offers, onEditClick }) {
-    super();
-    this.#point = point;
-    this.#destinations = destinations;
-    this.#offers = offers;
-    this.#handleEditClick = onEditClick;
-
-    this.element
-      .querySelector('.event__rollup-btn')
-      .addEventListener('click', this.#editClickHandler);
-  }
-
-  get template() {
-    return createPointTemplate({
-      point: this.#point,
-      destinations: this.#destinations,
-      offers: this.#offers,
-    });
-  }
-
-  #editClickHandler = (evt) => {
-    evt.preventDefault();
-    this.#handleEditClick();
-  };
-}
+export default createPointTemplate;
