+++ src/view/sort-view/sort-view.js
@@ -1,0 +1,8 @@
+import AbstractView from '../../framework/view/abstract-view';
+import createSortTemplate from './template';
+
+export default class SortView extends AbstractView {
+  get template() {
+    return createSortTemplate();
+  }
+}
